@@ -2,15 +2,14 @@
 // Copyright (c) PlaceholderCompany. All rights reserved.
 // </copyright>
 
-namespace Challenges.C20201001_20201007.CombinationSum
-{
-    using System.Collections.Generic;
+namespace Challenges.C20201001_20201007.CombinationSum;
 
-    public class Solution
+using System.Collections.Generic;
+
+public class Solution
+{
+    public IList<IList<int>> CombinationSum(int[] candidates, int target)
     {
-        public IList<IList<int>> CombinationSum(int[] candidates, int target)
-        {
-            return null;
-        }
+        return null;
     }
 }
